+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Added -n option and/or setting to_line to None to not send out email notifications.
 
 ### Changed
+- merge:  If no "to email" address is detected, do not send email notifications.
 - run_program:  Process -n option if detected in the command line arguments.
 - run_program: Process error messages from load_cfg function.
 - load_cfg:  Return list of error messages detected to calling function.