+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [1.0.1] - 2021-03-03
+
+
 ## [1.0.0] - 2021-02-23
 - General Release.
 - Added -n option and/or setting to_line to None to not send out email notifications.